Adiconando atualizacao de status ao vincular documentos

+++ src/ClientManager.Application/DocumentApplication.cs
@@ -1,18 +1,26 @@
 using ClientManager.Domain.Core.Responses;
 using ClientManager.Domain.Enums;
 using FluentValidation;
+using ClientManager.Application.Interfaces;
+using ClientManager.Domain.Core.Interfaces.Services;
 
 namespace ClientManager.Application;
 
 public class DocumentApplication : IDocumentApplication
 {
     private readonly IDocumentService _documentService;
+    private readonly ICustomerService _customerService;
     private readonly IFileValidator _fileValidator;
     private readonly IValidator<IFormFile> _fluentValidator;
 
-    public DocumentApplication(IDocumentService documentService, IFileValidator fileValidator, IValidator<IFormFile> fluentValidator)
+    public DocumentApplication(
+        IDocumentService documentService, 
+        ICustomerService customerService,
+        IFileValidator fileValidator, 
+        IValidator<IFormFile> fluentValidator)
     {
         _documentService = documentService;
+        _customerService = customerService;
         _fileValidator = fileValidator;
         _fluentValidator = fluentValidator;
     }
@@ -27,6 +35,10 @@ public class DocumentApplication : IDocumentApplication
         }
 
         var res = await _documentService.AttachDocumentAsync(customerId, file, type, expiryDate).ConfigureAwait(false);
+        
+        // Re-evaluate customer status
+        await ReevaluateCustomerStatusAsync(customerId).ConfigureAwait(false);
+
         return ServiceResponse<Guid>.Ok(res, "DocumentAttached");
     }
 
@@ -41,7 +53,17 @@ public class DocumentApplication : IDocumentApplication
 
     public async Task<ServiceResponse<string>> DeleteDocumentAsync(Guid documentId)
     {
+        var document = await _documentService.GetDocumentByIdAsync(documentId).ConfigureAwait(false);
+        if (document == null)
+            return ServiceResponse<string>.Fail("DocumentNotFound");
+
+        var customerId = document.CustomerId;
+
         await _documentService.DeleteDocumentAsync(documentId).ConfigureAwait(false);
+        
+        // Re-evaluate customer status
+        await ReevaluateCustomerStatusAsync(customerId).ConfigureAwait(false);
+
         return ServiceResponse<string>.Ok(documentId.ToString(), "DocumentRemoved");
     }
 
@@ -49,5 +71,16 @@ public class DocumentApplication : IDocumentApplication
     {
         var count = await _documentService.GetDocumentCountByCustomerIdAsync(customerId).ConfigureAwait(false);
         return ServiceResponse<int>.Ok(count);
+    }
+
+    private async Task ReevaluateCustomerStatusAsync(Guid customerId)
+    {
+        var customer = await _customerService.GetCustomerByIdAsync(customerId).ConfigureAwait(false);
+        if (customer != null)
+        {
+            var documents = await _documentService.GetDocumentsByCustomerIdAsync(customerId).ConfigureAwait(false);
+            customer.EvaluateVerificationStatus(documents);
+            await _customerService.UpdateCustomerAsync(customer).ConfigureAwait(false);
+        }
     }
 }
